+++ fts/src/lang/veclib/include/def_fill.h
@@ -97,11 +97,11 @@
  */
  
 #define DEFINE_FTS_VEC_FILL(name, t_scl, t_out, fun)\
- void fts_vec_ ## name(t_scl scl, t_out *out, long size)\
+ void fts_vec_ ## name(t_scl scl, t_out *out, int size)\
  {FTS_VEC_FILL(t_scl, t_out, fun)}
   
 #define DEFINE_FTS_VECX_FILL(name, t_scl, t_out, fun)\
- void fts_vecx_ ## name(t_scl scl, t_out *out, long size)\
+ void fts_vecx_ ## name(t_scl scl, t_out *out, int size)\
  {FTS_VECX_FILL(t_scl, t_out, fun)}
 
 #define DEFINE_FTL_VEC_FILL(name, t_scl, t_out, fun)\
@@ -126,10 +126,10 @@
  */
  
 #define DECLARE_FTS_VEC_FILL(name, t_scl, t_out, fun)\
- extern void fts_vec_ ## name(t_scl scl, t_out *out, long size);
+ extern void fts_vec_ ## name(t_scl scl, t_out *out, int size);
  
 #define DECLARE_FTS_VECX_FILL(name, t_scl, t_out, fun)\
- extern void fts_vecx_ ## name(t_scl scl, t_out *out, long size);
+ extern void fts_vecx_ ## name(t_scl scl, t_out *out, int size);
  
 #define DECLARE_FTL_VEC_FILL(name, t_scl, t_out, fun)\
  extern void ftl_vec_ ## name(FTL_ARG);
