+++ fts/src/lang/veclib/include/def_fun.h
@@ -102,11 +102,11 @@
  */
  
 #define DEFINE_FTS_VEC_FUN(name, t_in, t_out, fun)\
- void fts_vec_ ## name(t_in *in, t_out *out, long size)\
+ void fts_vec_ ## name(t_in *in, t_out *out, int size)\
  {FTS_VEC_FUN(t_in, t_out, fun)}
   
 #define DEFINE_FTS_VECX_FUN(name, t_in, t_out, fun)\
- void fts_vecx_ ## name(t_in *in, t_out *out, long size)\
+ void fts_vecx_ ## name(t_in *in, t_out *out, int size)\
  {FTS_VECX_FUN(t_in, t_out, fun)}
 
 #define DEFINE_FTL_VEC_FUN(name, t_in, t_out, fun)\
@@ -131,10 +131,10 @@
  */
  
 #define DECLARE_FTS_VEC_FUN(name, t_in, t_out, fun)\
- extern void fts_vec_ ## name(t_in *in, t_out *out, long size);
+ extern void fts_vec_ ## name(t_in *in, t_out *out, int size);
  
 #define DECLARE_FTS_VECX_FUN(name, t_in, t_out, fun)\
- extern void fts_vecx_ ## name(t_in *in, t_out *out, long size);
+ extern void fts_vecx_ ## name(t_in *in, t_out *out, int size);
  
 #define DECLARE_FTL_VEC_FUN(name, t_in, t_out, fun)\
  extern void ftl_vec_ ## name(FTL_ARG);
