+++ fts/src/lang/veclib/include/def_opscl.h
@@ -102,11 +102,11 @@
  */
  
 #define DEFINE_FTS_VEC_OPSCL(name, t_in, t_scl, t_out, op)\
- void fts_vec_ ## name(t_in *in, t_scl scl, t_out *out, long size)\
+ void fts_vec_ ## name(t_in *in, t_scl scl, t_out *out, int size)\
  {FTS_VEC_OPSCL(t_in, t_scl, t_out, op)}
   
 #define DEFINE_FTS_VECX_OPSCL(name, t_in, t_scl, t_out, op)\
- void fts_vecx_ ## name(t_in *in, t_scl scl, t_out *out, long size)\
+ void fts_vecx_ ## name(t_in *in, t_scl scl, t_out *out, int size)\
  {FTS_VECX_OPSCL(t_in, t_scl, t_out, op)}
 
 #define DEFINE_FTL_VEC_OPSCL(name, t_in, t_scl, t_out, op)\
@@ -131,10 +131,10 @@
  */
  
 #define DECLARE_FTS_VEC_OPSCL(name, t_in, t_scl, t_out, op)\
- extern void fts_vec_ ## name(t_in *in, t_scl scl, t_out *out, long size);
+ extern void fts_vec_ ## name(t_in *in, t_scl scl, t_out *out, int size);
  
 #define DECLARE_FTS_VECX_OPSCL(name, t_in, t_scl, t_out, op)\
- extern void fts_vecx_ ## name(t_in *in, t_scl scl, t_out *out, long size);
+ extern void fts_vecx_ ## name(t_in *in, t_scl scl, t_out *out, int size);
  
 #define DECLARE_FTL_VEC_OPSCL(name, t_in, t_scl, t_out, op)\
  extern void ftl_vec_ ## name(FTL_ARG);
