+++ fts/src/lang/veclib/include/def_set.h
@@ -96,11 +96,11 @@
  */
  
 #define DEFINE_FTS_VEC_SET(name, t_out, fun)\
- void fts_vec_ ## name(t_out *out, long size)\
+ void fts_vec_ ## name(t_out *out, int size)\
  {FTS_VEC_SET(t_out, fun)}
   
 #define DEFINE_FTS_VECX_SET(name, t_out, fun)\
- void fts_vecx_ ## name(t_out *out, long size)\
+ void fts_vecx_ ## name(t_out *out, int size)\
  {FTS_VECX_SET(t_out, fun)}
 
 #define DEFINE_FTL_VEC_SET(name, t_out, fun)\
@@ -125,10 +125,10 @@
  */
  
 #define DECLARE_FTS_VEC_SET(name, t_out, fun)\
- extern void fts_vec_ ## name(t_out *out, long size);
+ extern void fts_vec_ ## name(t_out *out, int size);
  
 #define DECLARE_FTS_VECX_SET(name, t_out, fun)\
- extern void fts_vecx_ ## name(t_out *out, long size);
+ extern void fts_vecx_ ## name(t_out *out, int size);
  
 #define DECLARE_FTL_VEC_SET(name, t_out, fun)\
  extern void ftl_vec_ ## name(FTL_ARG);
