+++ fts/src/lang/veclib/include/def_vecscl.h
@@ -103,11 +103,11 @@
  */
  
 #define DEFINE_FTS_VEC_VECSCL(name, t_in, t_ptr, op, init)\
- void fts_vec_ ## name(t_in *in, t_ptr *ptr, long size)\
+ void fts_vec_ ## name(t_in *in, t_ptr *ptr, int size)\
  {FTS_VEC_VECSCL(t_in, t_ptr, op, init)}
   
 #define DEFINE_FTS_VECX_VECSCL(name, t_in, t_ptr, op, init)\
- void fts_vecx_ ## name(t_in *in, t_ptr *ptr, long size)\
+ void fts_vecx_ ## name(t_in *in, t_ptr *ptr, int size)\
  {FTS_VECX_VECSCL(t_in, t_ptr, op, init)}
 
 #define DEFINE_FTL_VEC_VECSCL(name, t_in, t_ptr, op, init)\
@@ -132,10 +132,10 @@
  */
  
 #define DECLARE_FTS_VEC_VECSCL(name, t_in, t_ptr, op, init)\
- extern void fts_vec_ ## name(t_in *in, t_ptr *ptr, long size);
+ extern void fts_vec_ ## name(t_in *in, t_ptr *ptr, int size);
  
 #define DECLARE_FTS_VECX_VECSCL(name, t_in, t_ptr, op, init)\
- extern void fts_vecx_ ## name(t_in *in, t_ptr *ptr, long size);
+ extern void fts_vecx_ ## name(t_in *in, t_ptr *ptr, int size);
  
 #define DECLARE_FTL_VEC_VECSCL(name, t_in, t_ptr, op, init)\
  extern void ftl_vec_ ## name(FTL_ARG);
