+++ fts/src/lang/veclib/include/vec_cpy.h
@@ -8,14 +8,14 @@
  */
 
 DECLARE_VEC_FUN(FUN, (fcpy, float, float, FUN_cpy))
-DECLARE_VEC_FUN(FUN, (icpy, long, long, FUN_cpy))
+DECLARE_VEC_FUN(FUN, (icpy, int, int, FUN_cpy))
 DECLARE_VEC_FUN(CFUN, (ccpy, complex, complex, FUN_ccpy))
 
 DECLARE_VEC_FUN(FUN, (fcpyre, float, complex, FUN_fcpyre))
 DECLARE_VEC_FUN(FUN, (fcpyim, float, complex, FUN_fcpyim))
 
 DECLARE_VEC_FUN(FILL, (ffill, float, float, FUN_cpy))
-DECLARE_VEC_FUN(FILL, (ifill, long, long, FUN_cpy))
+DECLARE_VEC_FUN(FILL, (ifill, int, int, FUN_cpy))
 DECLARE_VEC_FUN(FILL, (cfill, complex, complex, FUN_ccpy))
 DECLARE_VEC_FUN(FILL, (ffillre, float, complex, FUN_fcpyre))
 DECLARE_VEC_FUN(FILL, (ffillim, float, complex, FUN_fcpyim))
@@ -24,14 +24,14 @@ DECLARE_VEC_FUN(FUN, (re, complex, float, FUN_re))
 DECLARE_VEC_FUN(FUN, (im, complex, float, FUN_im))
 
 DECLARE_VEC_FUN(SET, (fzero, float, FUN_zero))
-DECLARE_VEC_FUN(SET, (izero, long, FUN_zero))
+DECLARE_VEC_FUN(SET, (izero, int, FUN_zero))
 DECLARE_VEC_FUN(SET, (czero, complex, FUN_czero))
 
 /***************************************************
  *
- *		float/long conversion 
+ *		float/int conversion 
  *
  */
 
-DECLARE_VEC_FUN(FUN, (fcasti, float, long, FUN_cpy))
-DECLARE_VEC_FUN(FUN, (icastf, long, float, FUN_cpy))
+DECLARE_VEC_FUN(FUN, (fcasti, float, int, FUN_cpy))
+DECLARE_VEC_FUN(FUN, (icastf, int, float, FUN_cpy))
