+++ fts/src/lang/veclib/portable/vec_round.c
@@ -1,5 +1,5 @@
 #include "sys.h"
 #include "lang/veclib/include/vecdef.h"
 
-DEFINE_FTS_VEC_FUN(round, float, long, FUN_round)
-DEFINE_FTL_VEC_FUN(round, float, long, FUN_round)
+DEFINE_FTS_VEC_FUN(round, float, int, FUN_round)
+DEFINE_FTL_VEC_FUN(round, float, int, FUN_round)
